correct circular reference
