+++ src/Zikula/CoreInstallerBundle/Command/UpgradeCommand.php
@@ -31,7 +31,6 @@ use Zikula\Bundle\CoreInstallerBundle\Helper\ControllerHelper;
 use Zikula\Bundle\CoreInstallerBundle\Helper\MigrationHelper;
 use Zikula\Bundle\CoreInstallerBundle\Helper\StageHelper;
 use Zikula\Bundle\CoreInstallerBundle\Stage\Upgrade\AjaxUpgraderStage;
-use Zikula\Component\Wizard\StageContainerInterface;
 use Zikula\SettingsModule\Api\ApiInterface\LocaleApiInterface;
 
 class UpgradeCommand extends AbstractCoreInstallerCommand
@@ -69,9 +68,9 @@ class UpgradeCommand extends AbstractCoreInstallerCommand
     private $stageHelper;
 
     /**
-     * @var StageContainerInterface
+     * @var AjaxUpgraderStage
      */
-    private $stageContainer;
+    private $ajaxUpgraderStage;
 
     /**
      * @var array
@@ -91,7 +90,7 @@ class UpgradeCommand extends AbstractCoreInstallerCommand
         MigrationHelper $migrationHelper,
         LocaleApiInterface $localeApi,
         StageHelper $stageHelper,
-        StageContainerInterface $stageContainer,
+        AjaxUpgraderStage $ajaxUpgraderStage,
         TranslatorInterface $translator,
         ParameterBagInterface $params,
         string $installed
@@ -100,7 +99,7 @@ class UpgradeCommand extends AbstractCoreInstallerCommand
         $this->migrationHelper = $migrationHelper;
         $this->localeApi = $localeApi;
         $this->stageHelper = $stageHelper;
-        $this->stageContainer = $stageContainer;
+        $this->ajaxUpgraderStage = $ajaxUpgraderStage;
         $this->params = $params;
         $this->installed = $installed;
         parent::__construct($kernel, $translator);
@@ -185,8 +184,7 @@ class UpgradeCommand extends AbstractCoreInstallerCommand
         $yamlManager->setParameters($params);
 
         // upgrade!
-        $ajaxStage = $this->stageContainer->get(AjaxUpgraderStage::class);
-        $this->stageHelper->handleAjaxStage($ajaxStage, $io);
+        $this->stageHelper->handleAjaxStage($this->ajaxUpgraderStage, $io);
 
         $io->success($this->translator->trans('UPGRADE COMPLETE!'));
